Cmabios Edu

+++ src/main/java/com/app/ApiRestFul/services/AdminService.java
@@ -11,23 +11,32 @@ import java.util.Optional;
 
 @Service
 public class AdminService {
-	@Autowired
-	AdminRepository repository;
-        
-        
+
+    @Autowired
+    AdminRepository repository;
+
     public List<Admin> getAllAdmin() {
         List<Admin> admin = repository.findAll();
         return admin;
     }
 
-    public Admin getAdminById(String id) throws RecordNotFoundException {
-        Optional<Admin> client = repository.findById(id);
+    public Admin getAdminById(String id) throws RecordNotFoundException, IllegalArgumentException, NullPointerException {
+        if (id != null) {
+            try {
+                Optional<Admin> client = repository.findById(id);
 
-        if (client.isPresent()) {
-            return client.get();
+                if (client.isPresent()) {
+                    return client.get();
+                } else {
+                    throw new RecordNotFoundException("No client record exist for given id", id);
+                }
+            } catch (IllegalArgumentException e) {
+                throw new IllegalArgumentException(e);
+            }
         } else {
-            throw new RecordNotFoundException("No client record exist for given id", id);
+            throw new NullPointerException("Valor nulo");
         }
+
     }
 
     public Admin createAdmin(Admin admin) {
@@ -35,31 +44,47 @@ public class AdminService {
         return admin;
     }
 
-    public Admin UpdateAdmin(Admin entity) {
+    public Admin UpdateAdmin(Admin entity) throws RecordNotFoundException, IllegalArgumentException, NullPointerException {
         if (entity.getUser() != null) {
-            Optional<Admin> admin = repository.findById(entity.getUser());
+            try {
+                Optional<Admin> admin = repository.findById(entity.getUser());
 
-            if (admin.isPresent()) {
-                Admin newAdmin = admin.get();
-                newAdmin.setEmail(entity.getEmail());
-                newAdmin.setPassword(entity.getPassword());
- 
-                return newAdmin;
-            } else {
-                throw new RecordNotFoundException("Client not found",entity.getUser());
+                if (admin.isPresent()) {
+                    Admin newAdmin = admin.get();
+                    newAdmin.setEmail(entity.getEmail());
+                    newAdmin.setPassword(entity.getPassword());
+
+                    return newAdmin;
+                } else {
+                    throw new RecordNotFoundException("Client not found", entity.getUser());
+                }
+
+            } catch (IllegalArgumentException e) {
+                throw new IllegalArgumentException(e);
             }
+
         } else {
-            throw new RecordNotFoundException("Client not found",entity.getUser());
+            throw new NullPointerException("Valor Nulo");
         }
 
     }
 
-    public void deleteAdmin(String id) throws RecordNotFoundException {
-        Optional<Admin> Admin = repository.findById(id);
-        if (Admin.isPresent()) {
-            repository.deleteById(id);
+    public void deleteAdmin(String id) throws RecordNotFoundException, IllegalArgumentException, NullPointerException {
+
+        if (id != null) {
+            try {
+                Optional<Admin> Admin = repository.findById(id);
+                if (Admin.isPresent()) {
+                    repository.deleteById(id);
+                } else {
+                    throw new RecordNotFoundException("No client record exist for given id", id);
+                }
+            } catch (IllegalArgumentException e) {
+                throw new IllegalArgumentException(e);
+            }
         } else {
-            throw new RecordNotFoundException("No client record exist for given id", id);
+            throw new NullPointerException("Valor nulo");
         }
+
     }
 }
